django_backend/requirements.txt
Django==3.2.5
djangorestframework==3.12.4
django-cors-headers==3.7.0
requests==2.25.1
django-environ